Act on the D9 review: the modulus lost its unsigned wrap

+++ x/kdesigelSources.1.3/kdesigel/kdesigel/src/SIGEL_Simulation/SIG_DynaMechsCommandInterface.cpp
@@ -71,7 +71,16 @@ void SIGEL_Simulation::SIG_DynaMechsCommandInterface::moveDrive(int driveNo,
 
       int absoluteDriveNo = driveNo - static_cast< int >(minRegisterValue);
 
-      driveIndex = absoluteDriveNo % simulationData.drives.size();
+      // The modulus is UNSIGNED on purpose. Q2PtrVector::size() returned uint,
+      // so this promoted absoluteDriveNo and the result was always in
+      // [0, size). QList::size() is signed, which would hand a negative
+      // absoluteDriveNo straight through as a negative index -- into
+      // drives[] here and into driveForcesTimeAccounts[] below, which is a
+      // write. absoluteDriveNo goes negative whenever bitsPerRegister is 32,
+      // the SIG_LanguageParameters default: driveNo - (int)(-2^31) overflows.
+      // The wrap decides which actuator MOVE drives, so reproducing it is the
+      // point -- see PORTING.md section 9.
+      driveIndex = absoluteDriveNo % static_cast< uint >(simulationData.drives.size());
 
 
 #ifdef SIG_DEBUG
